@@ -69,7 +69,6 @@ import argparse
 import logging
 import sys
 from pathlib import Path
-from typing import Dict
 
 import coloredlogs
 import pandas as pd
@@ -96,7 +95,7 @@ MIN_PLANT_CAPACITY_MW: float = 5.0
 MAX_LOST_PLANTS_EIA: int = 50
 MAX_LOST_UTILS_EIA: int = 10
 
-PUDL_SETTINGS: Dict[str, str] = pudl.workspace.setup.get_defaults()
+PUDL_SETTINGS: dict[str, str] = pudl.workspace.setup.get_defaults()
 
 
 def parse_command_line(argv: str) -> argparse.Namespace:
