@@ -1,6 +1,6 @@
 """Field metadata."""
 from copy import deepcopy
-from typing import Any, Dict, Optional
+from typing import Any
 
 import pandas as pd
 from pytz import all_timezones
@@ -27,7 +27,7 @@ from pudl.metadata.labels import (
 )
 from pudl.metadata.sources import SOURCES
 
-FIELD_METADATA: Dict[str, Dict[str, Any]] = {
+FIELD_METADATA: dict[str, dict[str, Any]] = {
     "active": {
         "type": "boolean",
         "description": "Indicates whether or not the dataset has been pulled into PUDL by the extract transform load process.",
@@ -2016,7 +2016,7 @@ FIELD_METADATA: Dict[str, Dict[str, Any]] = {
 Keys are in alphabetical order.
 """
 
-FIELD_METADATA_BY_GROUP: Dict[str, Dict[str, Any]] = {
+FIELD_METADATA_BY_GROUP: dict[str, dict[str, Any]] = {
     "epacems": {
         "state": {"constraints": {"enum": EPACEMS_STATES}},
         "gross_load_mw": {
@@ -2079,7 +2079,7 @@ and has distinct metadata in those groups, this is the place to specify the
 override. Only those elements which should be overridden need to be specified.
 """
 
-FIELD_METADATA_BY_RESOURCE: Dict[str, Dict[str, Any]] = {
+FIELD_METADATA_BY_RESOURCE: dict[str, dict[str, Any]] = {
     "sector_consolidated_eia": {"code": {"type": "integer"}},
     "plants_steam_ferc1": {
         "plant_type": {
@@ -2103,11 +2103,11 @@ FIELD_METADATA_BY_RESOURCE: Dict[str, Dict[str, Any]] = {
 
 
 def get_pudl_dtypes(
-    group: Optional[str] = None,
-    field_meta: Optional[Dict[str, Any]] = FIELD_METADATA,
-    field_meta_by_group: Optional[Dict[str, Any]] = FIELD_METADATA_BY_GROUP,
-    dtype_map: Optional[Dict[str, Any]] = FIELD_DTYPES_PANDAS,
-) -> Dict[str, Any]:
+    group: str | None = None,
+    field_meta: dict[str, Any] | None = FIELD_METADATA,
+    field_meta_by_group: dict[str, Any] | None = FIELD_METADATA_BY_GROUP,
+    dtype_map: dict[str, Any] | None = FIELD_DTYPES_PANDAS,
+) -> dict[str, Any]:
     """Compile a dictionary of field dtypes, applying group overrides.
 
     Args:
@@ -2136,9 +2136,9 @@ def get_pudl_dtypes(
 
 def apply_pudl_dtypes(
     df: pd.DataFrame,
-    group: Optional[str] = None,
-    field_meta: Optional[Dict[str, Any]] = FIELD_METADATA,
-    field_meta_by_group: Optional[Dict[str, Any]] = FIELD_METADATA_BY_GROUP,
+    group: str | None = None,
+    field_meta: dict[str, Any] | None = FIELD_METADATA,
+    field_meta_by_group: dict[str, Any] | None = FIELD_METADATA_BY_GROUP,
 ) -> pd.DataFrame:
     """Apply dtypes to those columns in a dataframe that have PUDL types defined.
 
