@@ -1,7 +1,4 @@
 """Tests for timeseries anomalies detection and imputation."""
-
-from typing import Dict, Union
-
 import numpy as np
 import pandas as pd
 import pytest
@@ -33,8 +30,6 @@ AK_FIPS = {"name": "Alaska", "code": "AK", "fips": "02"}
         pytest.param(None, {}, marks=pytest.mark.xfail),
     ],
 )
-def test_lookup_state(
-    state: Union[str, int], expected: Dict[str, Union[str, int]]
-) -> None:
+def test_lookup_state(state: str | int, expected: dict[str, str | int]) -> None:
     """Check that various kinds of state lookups work."""
     assert lookup_state(state) == expected
