@@ -28,7 +28,7 @@ def test_datasette_metadata_script(script_runner, pudl_settings_fixture):
     logger.info("Parsing generated metadata using datasette utils.")
     metadata_json = json.dumps(yaml.safe_load(metadata_yml.open()))
     parsed_metadata = datasette.utils.parse_metadata(metadata_json)
-    assert set(parsed_metadata["databases"]) == set(["pudl", "ferc1"])
+    assert set(parsed_metadata["databases"]) == {"pudl", "ferc1"}
     assert parsed_metadata["license"] == "CC-BY-4.0"
     assert (
         parsed_metadata["databases"]["pudl"]["source_url"]
